cluster/requirements.txt
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
grpcio
huggingface-hub
nltk
protobuf==3.20
transformers
torch
python-dotenv
scipy
ffmpeg-python